Select in other frame
The strokes in Plover are too fast for Emacs.  It gets confused after
M-o.  So, I put a bunch of NULL strokes to kill time until Emacs is
ready.  This might be better solved by having Plover directly send
Emacs commands.  Or maybe not–this seems to work quite well.